refactor: 참가자가 Money를 인스턴스 필드로 가지게 수정
