--- src/main/java/blackjack/domain/participant/DealerMoney.java
@@ -1,17 +1,0 @@
-package blackjack.domain.participant;
-
-public class DealerMoney {
-    private int money;
-
-    public DealerMoney() {
-        money = 0;
-    }
-
-    public void calculateByOpponentProfit(final int opponentProfit) {
-        money = money - opponentProfit;
-    }
-
-    public int getMoney() {
-        return money;
-    }
-}
